Blogs refactoring (#747)
Refactor blogs module, add retries and logging to publish/delete s3 tasks.
