+++ server/liveblog/blogs/blogslist_assets/main.js
