+++ server/liveblog/blogs/embeds.py
@@ -8,41 +8,36 @@
 # For the full copyright and license information, please see the
 # AUTHORS and LICENSE files distributed with this source code, or
 # at https://www.sourcefabric.org/superdesk/license
-from superdesk.errors import SuperdeskApiError
 
-"""Embed module"""
-import superdesk
-from flask import render_template, json, request, current_app as app
-from eve.io.mongo import MongoJSONEncoder
-from superdesk import get_resource_service
-from liveblog.themes import UnknownTheme, ASSETS_DIR as THEMES_ASSETS_DIR
-from flask import url_for
-import io
-import os
 import json
 import logging
+import os
+
+import superdesk
+from eve.io.mongo import MongoJSONEncoder
+from flask import current_app as app
+from flask import json, render_template, request, url_for
+from liveblog.themes import ASSETS_DIR as THEMES_ASSETS_DIR
+from liveblog.themes import UnknownTheme
+from superdesk import get_resource_service
+from superdesk.errors import SuperdeskApiError
+
+from .utils import is_relative_to_current_folder
 
 logger = logging.getLogger('superdesk')
+embed_blueprint = superdesk.Blueprint('embed_liveblog', __name__, template_folder='templates')
 THEMES_DIRECTORY = os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), os.pardir, 'themes'))
-bp = superdesk.Blueprint('embed_liveblog', __name__, template_folder='templates')
-
-
-class MediaStorageUnsupportedForBlogPublishing(Exception):
-    pass
-
-
-def is_relative_to_current_folder(url):
-    return not (url.startswith('/') or url.startswith('http://') or url.startswith('https://'))
 
 
 def collect_theme_assets(theme, assets=None, template=None):
     assets = assets or {'scripts': [], 'styles': [], 'devScripts': [], 'devStyles': []}
-    # load the template
+    # Load the template.
     if not template:
         template_file_name = os.path.join(THEMES_DIRECTORY, THEMES_ASSETS_DIR, theme['name'], 'template.html')
         if os.path.isfile(template_file_name):
             template = open(template_file_name, encoding='utf-8').read()
-    # add assets from parent theme
+
+    # Add assets from parent theme.
     if theme.get('extends', None):
         parent_theme = get_resource_service('themes').find_one(req=None, name=theme.get('extends'))
         if parent_theme:
@@ -52,7 +47,8 @@ def collect_theme_assets(theme, assets=None, template=None):
                 % (theme.get('name'), theme.get('extends'))
             logger.info(error_message)
             raise UnknownTheme(error_message)
-    # add assets from theme
+
+    # Add assets from theme.
     for asset_type in ('scripts', 'styles', 'devScripts', 'devStyles'):
         theme_folder = theme['name']
         for url in theme.get(asset_type, []):
@@ -62,70 +58,48 @@ def collect_theme_assets(theme, assets=None, template=None):
                 else:
                     url = url_for('themes_assets.static', filename=os.path.join(theme_folder, url), _external=False)
             assets[asset_type].append(url)
+
     return assets, template
 
 
-def get_file_path(blog_id):
-    return 'blogs/%s/index.html' % (blog_id)
-
-
-def check_media_storage():
-    if type(app.media).__name__ is not 'AmazonMediaStorage':
-        raise MediaStorageUnsupportedForBlogPublishing()
-
-
-def publish_embed(blog_id, api_host=None, theme=None):
-    html = embed(blog_id, api_host, theme)
-    check_media_storage()
-    file_path = get_file_path(blog_id)
-    # remove existing
-    app.media.delete(app.media.media_id(file_path, version=False))
-    # upload
-    file_id = app.media.put(io.BytesIO(bytes(html, 'utf-8')),
-                            filename=file_path,
-                            content_type='text/html',
-                            version=False)
-    return superdesk.upload.url_for_media(file_id)
-
-
-def delete_embed(blog_id):
-    check_media_storage()
-    file_path = get_file_path(blog_id)
-    # remove existing
-    app.media.delete(file_path)
-
-
-@bp.route('/embed/<blog_id>')
+@embed_blueprint.route('/embed/<blog_id>')
 def embed(blog_id, api_host=None, theme=None):
     api_host = api_host or request.url_root
     blog = get_resource_service('client_blogs').find_one(req=None, _id=blog_id)
     if not blog:
         return 'blog not found', 404
-    # retrieve picture url from relationship
+
+    # Retrieve picture url from relationship.
     if blog.get('picture', None):
         blog['picture'] = get_resource_service('archive').find_one(req=None, _id=blog['picture'])
-    # retrieve the wanted theme and add it to blog['theme'] if is not the registered one
+
+    # Retrieve the wanted theme and add it to blog['theme'] if is not the registered one.
     try:
         theme_name = request.args.get('theme', theme)
     except RuntimeError:
-        # this method can be called outside from a request context
+        # This method can be called outside from a request context.
         theme_name = theme
+
     theme = get_resource_service('themes').find_one(req=None, name=blog['blog_preferences'].get('theme'))
     if theme is None and theme_name is None:
         raise SuperdeskApiError.badRequestError(
             message='You will be able to access the embed after you register the themes')
-    # if a theme is provided, overwrite the default theme
+
+    # If a theme is provided, overwrite the default theme.
     if theme_name:
         theme_package = os.path.join(THEMES_DIRECTORY, THEMES_ASSETS_DIR, theme_name, 'theme.json')
         theme = json.loads(open(theme_package).read())
+
     try:
         assets, template_file = collect_theme_assets(theme)
     except UnknownTheme as e:
         return str(e), 500
+
     if not template_file:
         logger.error('Template file not found for theme "%s". Theme: %s' % (theme.get('name'), theme))
         return 'Template file not found', 500
-    # compute the assets root
+
+    # Compute the assets root.
     if theme.get('public_url', False):
         assets_root = theme.get('public_url')
     else:
@@ -144,7 +118,7 @@ def embed(blog_id, api_host=None, theme=None):
     return render_template('embed.html', **scope)
 
 
-@bp.route('/embed/<blog_id>/overview')
+@embed_blueprint.route('/embed/<blog_id>/overview')
 def embed_overview(blog_id, api_host=None):
     ''' Show a theme with all the available themes in different iframes '''
     blog = get_resource_service('client_blogs').find_one(req=None, _id=blog_id)
@@ -157,13 +131,11 @@ def embed_overview(blog_id, api_host=None):
     return render_template('iframe-for-every-themes.html', **scope)
 
 
-@bp.app_template_filter('tojson')
+@embed_blueprint.app_template_filter('tojson')
 def tojson(obj):
     return json.dumps(obj, cls=MongoJSONEncoder)
 
 
-@bp.app_template_filter('is_relative_to_current_folder')
+@embed_blueprint.app_template_filter('is_relative_to_current_folder')
 def is_relative_to_current_folder_filter(s):
     return is_relative_to_current_folder(s)
-
-# EOF
