+++ server/liveblog/blogs/exceptions.py
@@ -1,0 +1,2 @@
+class MediaStorageUnsupportedForBlogPublishing(Exception):
+    pass
